Completed documentations for GodotEx basics

+++ src/GodotEx/src/Classes/BasisExtensions.cs
@@ -1,4 +1,4 @@
-using DotnetEx.Maths;
+using DotEx.Maths;
 using Godot;
 
 namespace GodotEx;
@@ -13,7 +13,7 @@ public static class BasisExtensions {
     /// <param name="basis">Basis to trim.</param>
     /// <param name="precision">Precision to use.</param>
     /// <returns>Trimmed basis.</returns>
-    public static Basis Trim(this Basis basis, float precision = MathDef.DEFAULT_PRECISION) {
+    public static Basis Trim(this Basis basis, float precision = (float)MathDef.DEFAULT_PRECISION) {
         return new(basis.X.Trim(precision), basis.Y.Trim(precision), basis.Z.Trim(precision));
     }
 }
